refactor: 디렉토리 구조 깊이 조정
