+++ src/core/index.ts
@@ -1,2 +1,4 @@
 export * from './render';
 export * from './constants';
+export * from './jsx-runtime';
+export * from './applyRef';
