+++ src/core/jsx-runtime.ts
@@ -1,11 +1,10 @@
-// src/jsx-runtime/jsx-runtime.ts
-import { FRAGMENT, TEXT_ELEMENT } from '../core/constants';
 import {
 	type FragmentVNode,
 	TextVNode,
 	type VNode,
 	type VNodeChildren,
 } from '../types/base.types';
+import { FRAGMENT, TEXT_ELEMENT } from './constants';
 
 function wrapChild(child: VNode | string | number): VNode {
 	if (typeof child === 'string' || typeof child === 'number') {
