--- src/jsx-runtime/index.ts
@@ -1,1 +1,0 @@
-export * from './jsx-runtime';
